+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ### Changes ###
 
-  * Decided to remove the Vehicle Bar. Sorry for the confusion.
+  * Adjust AFK/DND Colors on my NameHover to reflect ElvUI Changes.
+  * Updated the code from HealPrediction. Credits: Windtools
+  * Added new Texture 4Pixel.
+  * Added more options to change the Objective Tracker skin. Credits: Windtools
+  * Added an Item Level String to the Guild News. Credits: Windtools